requirements.txt
web3>=5.31.1
python-dotenv>=0.20.0
requests>=2.28.0
pandas>=1.4.2